+++ src/engine/renderer/tr_init.cpp
@@ -23,263 +23,9 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 // tr_init.c -- functions that are not called every frame
 #include "tr_local.h"
 
-	glconfig_t  glConfig;
-	glconfig2_t glConfig2;
-
-	glstate_t   glState;
-
 	float       displayAspect = 0.0f;
 
 	static void GfxInfo_f();
-
-	cvar_t      *r_glMajorVersion;
-	cvar_t      *r_glMinorVersion;
-	cvar_t      *r_glProfile;
-	cvar_t      *r_glDebugProfile;
-	cvar_t      *r_glDebugMode;
-	cvar_t      *r_glAllowSoftware;
-
-	cvar_t      *r_verbose;
-	cvar_t      *r_ignore;
-
-	cvar_t      *r_znear;
-	cvar_t      *r_zfar;
-
-	cvar_t      *r_smp;
-	cvar_t      *r_showSmp;
-	cvar_t      *r_skipBackEnd;
-	cvar_t      *r_skipLightBuffer;
-
-	cvar_t      *r_measureOverdraw;
-
-	cvar_t      *r_fastsky;
-	cvar_t      *r_drawSun;
-
-	cvar_t      *r_lodBias;
-	cvar_t      *r_lodScale;
-	cvar_t      *r_lodTest;
-
-	cvar_t      *r_norefresh;
-	cvar_t      *r_drawentities;
-	cvar_t      *r_drawworld;
-	cvar_t      *r_drawpolies;
-	cvar_t      *r_speeds;
-	cvar_t      *r_novis;
-	cvar_t      *r_nocull;
-	cvar_t      *r_facePlaneCull;
-	cvar_t      *r_showcluster;
-	cvar_t      *r_nocurves;
-	cvar_t      *r_lightScissors;
-	cvar_t      *r_noLightVisCull;
-	cvar_t      *r_noInteractionSort;
-	cvar_t      *r_dynamicLight;
-	cvar_t      *r_staticLight;
-	cvar_t      *r_dynamicLightCastShadows;
-	cvar_t      *r_precomputedLighting;
-	cvar_t      *r_vertexLighting;
-	cvar_t      *r_exportTextures;
-	cvar_t      *r_heatHaze;
-	cvar_t      *r_noMarksOnTrisurfs;
-	cvar_t      *r_lazyShaders;
-
-	cvar_t      *r_ext_occlusion_query;
-	cvar_t      *r_ext_draw_buffers;
-	cvar_t      *r_ext_vertex_array_object;
-	cvar_t      *r_ext_half_float_pixel;
-	cvar_t      *r_ext_texture_float;
-	cvar_t      *r_ext_texture_integer;
-	cvar_t      *r_ext_texture_rg;
-	cvar_t      *r_ext_texture_filter_anisotropic;
-	cvar_t      *r_ext_gpu_shader4;
-	cvar_t      *r_arb_buffer_storage;
-	cvar_t      *r_arb_map_buffer_range;
-	cvar_t      *r_arb_sync;
-	cvar_t      *r_arb_uniform_buffer_object;
-	cvar_t      *r_arb_texture_gather;
-	cvar_t      *r_arb_gpu_shader5;
-
-	cvar_t      *r_checkGLErrors;
-	cvar_t      *r_logFile;
-
-	cvar_t      *r_stencilbits;
-	cvar_t      *r_depthbits;
-	cvar_t      *r_colorbits;
-	cvar_t      *r_alphabits;
-	cvar_t      *r_ext_multisample;
-
-	cvar_t      *r_drawBuffer;
-	cvar_t      *r_shadows;
-	cvar_t      *r_softShadows;
-	cvar_t      *r_softShadowsPP;
-	cvar_t      *r_shadowBlur;
-
-	cvar_t      *r_shadowMapQuality;
-	cvar_t      *r_shadowMapSizeUltra;
-	cvar_t      *r_shadowMapSizeVeryHigh;
-	cvar_t      *r_shadowMapSizeHigh;
-	cvar_t      *r_shadowMapSizeMedium;
-	cvar_t      *r_shadowMapSizeLow;
-
-	cvar_t      *r_shadowMapSizeSunUltra;
-	cvar_t      *r_shadowMapSizeSunVeryHigh;
-	cvar_t      *r_shadowMapSizeSunHigh;
-	cvar_t      *r_shadowMapSizeSunMedium;
-	cvar_t      *r_shadowMapSizeSunLow;
-
-	cvar_t      *r_shadowOffsetFactor;
-	cvar_t      *r_shadowOffsetUnits;
-	cvar_t      *r_shadowLodBias;
-	cvar_t      *r_shadowLodScale;
-	cvar_t      *r_noShadowPyramids;
-	cvar_t      *r_cullShadowPyramidFaces;
-	cvar_t      *r_cullShadowPyramidCurves;
-	cvar_t      *r_cullShadowPyramidTriangles;
-	cvar_t      *r_debugShadowMaps;
-	cvar_t      *r_noShadowFrustums;
-	cvar_t      *r_noLightFrustums;
-	cvar_t      *r_shadowMapLinearFilter;
-	cvar_t      *r_lightBleedReduction;
-	cvar_t      *r_overDarkeningFactor;
-	cvar_t      *r_shadowMapDepthScale;
-	cvar_t      *r_parallelShadowSplits;
-	cvar_t      *r_parallelShadowSplitWeight;
-
-	cvar_t      *r_mode;
-	cvar_t      *r_nobind;
-	cvar_t      *r_singleShader;
-	cvar_t      *r_colorMipLevels;
-	cvar_t      *r_picMip;
-	cvar_t      *r_imageMaxDimension;
-	cvar_t      *r_ignoreMaterialMinDimension;
-	cvar_t      *r_ignoreMaterialMaxDimension;
-	cvar_t      *r_replaceMaterialMinDimensionIfPresentWithMaxDimension;
-	cvar_t      *r_finish;
-	cvar_t      *r_clear;
-	cvar_t      *r_swapInterval;
-	cvar_t      *r_textureMode;
-	cvar_t      *r_offsetFactor;
-	cvar_t      *r_offsetUnits;
-
-	cvar_t      *r_physicalMapping;
-	cvar_t      *r_specularExponentMin;
-	cvar_t      *r_specularExponentMax;
-	cvar_t      *r_specularScale;
-	cvar_t      *r_specularMapping;
-	cvar_t      *r_deluxeMapping;
-	cvar_t      *r_normalScale;
-	cvar_t      *r_normalMapping;
-	cvar_t      *r_liquidMapping;
-	cvar_t      *r_highQualityNormalMapping;
-	cvar_t      *r_reliefDepthScale;
-	cvar_t      *r_reliefMapping;
-	cvar_t      *r_glowMapping;
-	cvar_t      *r_reflectionMapping;
-
-	cvar_t      *r_wrapAroundLighting;
-	cvar_t      *r_halfLambertLighting;
-	cvar_t      *r_rimLighting;
-	cvar_t      *r_rimExponent;
-	cvar_t      *r_gamma;
-	cvar_t      *r_lockpvs;
-	cvar_t      *r_noportals;
-	cvar_t      *r_portalOnly;
-	cvar_t      *r_portalSky;
-	cvar_t      *r_max_portal_levels;
-
-	cvar_t      *r_subdivisions;
-	cvar_t      *r_stitchCurves;
-
-	cvar_t      *r_noBorder;
-	cvar_t      *r_fullscreen;
-
-	cvar_t      *r_customwidth;
-	cvar_t      *r_customheight;
-
-	cvar_t      *r_debugSurface;
-	cvar_t      *r_simpleMipMaps;
-
-	cvar_t      *r_showImages;
-
-	cvar_t      *r_forceFog;
-	cvar_t      *r_wolfFog;
-	cvar_t      *r_noFog;
-
-	cvar_t      *r_forceAmbient;
-	cvar_t      *r_ambientScale;
-	cvar_t      *r_lightScale;
-	cvar_t      *r_debugSort;
-	cvar_t      *r_printShaders;
-
-	cvar_t      *r_maxPolys;
-	cvar_t      *r_maxPolyVerts;
-
-	cvar_t      *r_showTris;
-	cvar_t      *r_showSky;
-	cvar_t      *r_showShadowVolumes;
-	cvar_t      *r_showShadowLod;
-	cvar_t      *r_showShadowMaps;
-	cvar_t      *r_showSkeleton;
-	cvar_t      *r_showEntityTransforms;
-	cvar_t      *r_showLightTransforms;
-	cvar_t      *r_showLightInteractions;
-	cvar_t      *r_showLightScissors;
-	cvar_t      *r_showLightBatches;
-	cvar_t      *r_showLightGrid;
-	cvar_t      *r_showLightTiles;
-	cvar_t      *r_showBatches;
-	cvar_t      *r_showLightMaps;
-	cvar_t      *r_showDeluxeMaps;
-	cvar_t      *r_showNormalMaps;
-	cvar_t      *r_showMaterialMaps;
-	cvar_t      *r_showAreaPortals;
-	cvar_t      *r_showCubeProbes;
-	cvar_t      *r_showBspNodes;
-	cvar_t      *r_showParallelShadowSplits;
-	cvar_t      *r_showDecalProjectors;
-
-	cvar_t      *r_vboFaces;
-	cvar_t      *r_vboCurves;
-	cvar_t      *r_vboTriangles;
-	cvar_t      *r_vboShadows;
-	cvar_t      *r_vboLighting;
-	cvar_t      *r_vboModels;
-	cvar_t      *r_vboVertexSkinning;
-	cvar_t      *r_vboDeformVertexes;
-
-	cvar_t      *r_mergeLeafSurfaces;
-
-	cvar_t      *r_bloom;
-	cvar_t      *r_bloomBlur;
-	cvar_t      *r_bloomPasses;
-	cvar_t      *r_FXAA;
-	cvar_t      *r_ssao;
-
-	cvar_t      *r_evsmPostProcess;
-
-	cvar_t      *r_fontScale;
-
-	void AssertCvarRange( cvar_t *cv, float minVal, float maxVal, bool shouldBeIntegral )
-	{
-		if ( shouldBeIntegral )
-		{
-			if ( cv->value != static_cast<float>(cv->integer) )
-			{
-				Log::Warn("cvar '%s' must be integral (%f)", cv->name, cv->value );
-				ri.Cvar_Set( cv->name, va( "%d", cv->integer ) );
-			}
-		}
-
-		if ( cv->value < minVal )
-		{
-			Log::Warn("cvar '%s' out of range (%g < %g)", cv->name, cv->value, minVal );
-			ri.Cvar_Set( cv->name, va( "%f", minVal ) );
-		}
-		else if ( cv->value > maxVal )
-		{
-			Log::Warn("cvar '%s' out of range (%g > %g)", cv->name, cv->value, maxVal );
-			ri.Cvar_Set( cv->name, va( "%f", maxVal ) );
-		}
-	}
 
 	/*
 	** InitOpenGL
@@ -365,69 +111,6 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 		GL_CheckErrors();
 
 		return true;
-	}
-
-	/*
-	==================
-	GL_CheckErrors
-
-	Must not be called while the backend rendering thread is running
-	==================
-	*/
-	void GL_CheckErrors_( const char *fileName, int line )
-	{
-		int  err;
-		char s[ 128 ];
-
-		if ( !checkGLErrors() )
-		{
-			return;
-		}
-
-		while ( ( err = glGetError() ) != GL_NO_ERROR )
-		{
-			switch ( err )
-			{
-				case GL_INVALID_ENUM:
-					strcpy( s, "GL_INVALID_ENUM" );
-					break;
-
-				case GL_INVALID_VALUE:
-					strcpy( s, "GL_INVALID_VALUE" );
-					break;
-
-				case GL_INVALID_OPERATION:
-					strcpy( s, "GL_INVALID_OPERATION" );
-					break;
-
-				case GL_STACK_OVERFLOW:
-					strcpy( s, "GL_STACK_OVERFLOW" );
-					break;
-
-				case GL_STACK_UNDERFLOW:
-					strcpy( s, "GL_STACK_UNDERFLOW" );
-					break;
-
-				case GL_OUT_OF_MEMORY:
-					strcpy( s, "GL_OUT_OF_MEMORY" );
-					break;
-
-				case GL_TABLE_TOO_LARGE:
-					strcpy( s, "GL_TABLE_TOO_LARGE" );
-					break;
-
-				case GL_INVALID_FRAMEBUFFER_OPERATION:
-					strcpy( s, "GL_INVALID_FRAMEBUFFER_OPERATION" );
-					break;
-
-				default:
-					Com_sprintf( s, sizeof( s ), "0x%X", err );
-					break;
-			}
-			// Pre-format the string so that each callsite counts separately for log suppression
-			std::string error = Str::Format("OpenGL error %s detected at %s:%d", s, fileName, line);
-			Log::Warn(error);
-		}
 	}
 
 	/*
@@ -1027,282 +710,17 @@ ScreenshotCmd screenshotPNGRegistration("screenshotPNG", ssFormat_t::SSF_PNG, "p
 	*/
 	void R_Register()
 	{
-		// OpenGL context selection
-		r_glMajorVersion = ri.Cvar_Get( "r_glMajorVersion", "", CVAR_LATCH );
-		r_glMinorVersion = ri.Cvar_Get( "r_glMinorVersion", "", CVAR_LATCH );
-		r_glProfile = ri.Cvar_Get( "r_glProfile", "", CVAR_LATCH );
-		r_glDebugProfile = ri.Cvar_Get( "r_glDebugProfile", "", CVAR_LATCH );
-		r_glDebugMode = ri.Cvar_Get( "r_glDebugMode", "0", CVAR_CHEAT );
-		r_glAllowSoftware = ri.Cvar_Get( "r_glAllowSoftware", "0", CVAR_LATCH );
-
-		// latched and archived variables
-		r_ext_occlusion_query = ri.Cvar_Get( "r_ext_occlusion_query", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_ext_draw_buffers = ri.Cvar_Get( "r_ext_draw_buffers", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_ext_vertex_array_object = ri.Cvar_Get( "r_ext_vertex_array_object", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_ext_half_float_pixel = ri.Cvar_Get( "r_ext_half_float_pixel", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_ext_texture_float = ri.Cvar_Get( "r_ext_texture_float", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_ext_texture_integer = ri.Cvar_Get( "r_ext_texture_integer", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_ext_texture_rg = ri.Cvar_Get( "r_ext_texture_rg", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_ext_texture_filter_anisotropic = ri.Cvar_Get( "r_ext_texture_filter_anisotropic", "4",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_ext_gpu_shader4 = ri.Cvar_Get( "r_ext_gpu_shader4", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_arb_buffer_storage = ri.Cvar_Get( "r_arb_buffer_storage", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_arb_map_buffer_range = ri.Cvar_Get( "r_arb_map_buffer_range", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_arb_sync = ri.Cvar_Get( "r_arb_sync", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_arb_uniform_buffer_object = ri.Cvar_Get( "r_arb_uniform_buffer_object", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_arb_texture_gather = ri.Cvar_Get( "r_arb_texture_gather", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_arb_gpu_shader5 = ri.Cvar_Get( "r_arb_gpu_shader5", "1", CVAR_CHEAT | CVAR_LATCH );
-
-		r_picMip = ri.Cvar_Get( "r_picMip", "0",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_imageMaxDimension = ri.Cvar_Get( "r_imageMaxDimension", "0",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_ignoreMaterialMinDimension = ri.Cvar_Get( "r_ignoreMaterialMinDimension", "0",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_ignoreMaterialMaxDimension = ri.Cvar_Get( "r_ignoreMaterialMaxDimension", "0",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_replaceMaterialMinDimensionIfPresentWithMaxDimension
-			= ri.Cvar_Get( "r_replaceMaterialMinDimensionIfPresentWithMaxDimension", "0",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_colorMipLevels = ri.Cvar_Get( "r_colorMipLevels", "0", CVAR_LATCH );
-		r_colorbits = ri.Cvar_Get( "r_colorbits", "0",  CVAR_LATCH );
-		r_alphabits = ri.Cvar_Get( "r_alphabits", "0",  CVAR_LATCH );
-		r_stencilbits = ri.Cvar_Get( "r_stencilbits", "8",  CVAR_LATCH );
-		r_depthbits = ri.Cvar_Get( "r_depthbits", "0",  CVAR_LATCH );
-		r_ext_multisample = ri.Cvar_Get( "r_ext_multisample", "0",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_mode = ri.Cvar_Get( "r_mode", "-2", CVAR_LATCH | CVAR_ARCHIVE );
-		r_noBorder = ri.Cvar_Get( "r_noBorder", "0", CVAR_ARCHIVE );
-		r_fullscreen = ri.Cvar_Get( "r_fullscreen", "1", CVAR_ARCHIVE );
-		r_customwidth = ri.Cvar_Get( "r_customwidth", "1600", CVAR_LATCH | CVAR_ARCHIVE );
-		r_customheight = ri.Cvar_Get( "r_customheight", "1024", CVAR_LATCH | CVAR_ARCHIVE );
-		r_simpleMipMaps = ri.Cvar_Get( "r_simpleMipMaps", "0", CVAR_LATCH );
-		r_subdivisions = ri.Cvar_Get( "r_subdivisions", "4", CVAR_LATCH );
-		r_dynamicLightCastShadows = ri.Cvar_Get( "r_dynamicLightCastShadows", "1", 0 );
-		r_precomputedLighting = ri.Cvar_Get( "r_precomputedLighting", "1", CVAR_LATCH );
-		r_vertexLighting = ri.Cvar_Get( "r_vertexLighting", "0", CVAR_LATCH | CVAR_ARCHIVE );
-		r_exportTextures = ri.Cvar_Get( "r_exportTextures", "0", 0 );
-		r_heatHaze = ri.Cvar_Get( "r_heatHaze", "1", CVAR_LATCH | CVAR_ARCHIVE );
-		r_noMarksOnTrisurfs = ri.Cvar_Get( "r_noMarksOnTrisurfs", "1", CVAR_CHEAT );
-		r_lazyShaders = ri.Cvar_Get( "r_lazyShaders", "0", 0 );
-
-		r_forceFog = ri.Cvar_Get( "r_forceFog", "0", CVAR_CHEAT /* | CVAR_LATCH */ );
-		AssertCvarRange( r_forceFog, 0.0f, 1.0f, false );
-		r_wolfFog = ri.Cvar_Get( "r_wolfFog", "1", CVAR_CHEAT );
-		r_noFog = ri.Cvar_Get( "r_noFog", "0", CVAR_CHEAT );
-
-		r_forceAmbient = ri.Cvar_Get( "r_forceAmbient", "0.125",  CVAR_LATCH );
-		AssertCvarRange( r_forceAmbient, 0.0f, 0.3f, false );
-
-		r_smp = ri.Cvar_Get( "r_smp", "0",  CVAR_LATCH );
-
-		// temporary latched variables that can only change over a restart
-		r_singleShader = ri.Cvar_Get( "r_singleShader", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_stitchCurves = ri.Cvar_Get( "r_stitchCurves", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_debugShadowMaps = ri.Cvar_Get( "r_debugShadowMaps", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_shadowMapLinearFilter = ri.Cvar_Get( "r_shadowMapLinearFilter", "1", CVAR_CHEAT | CVAR_LATCH );
-		r_lightBleedReduction = ri.Cvar_Get( "r_lightBleedReduction", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_overDarkeningFactor = ri.Cvar_Get( "r_overDarkeningFactor", "30.0", CVAR_CHEAT | CVAR_LATCH );
-		r_shadowMapDepthScale = ri.Cvar_Get( "r_shadowMapDepthScale", "1.41", CVAR_CHEAT | CVAR_LATCH );
-
-		r_parallelShadowSplitWeight = ri.Cvar_Get( "r_parallelShadowSplitWeight", "0.9", CVAR_CHEAT );
-		r_parallelShadowSplits = ri.Cvar_Get( "r_parallelShadowSplits", "2", CVAR_CHEAT | CVAR_LATCH );
-		AssertCvarRange( r_parallelShadowSplits, 0, MAX_SHADOWMAPS - 1, true );
-
-		// archived variables that can change at any time
-		r_lodBias = ri.Cvar_Get( "r_lodBias", "0", 0 );
-		r_znear = ri.Cvar_Get( "r_znear", "3", CVAR_CHEAT );
-		r_zfar = ri.Cvar_Get( "r_zfar", "0", CVAR_CHEAT );
-		r_checkGLErrors = ri.Cvar_Get( "r_checkGLErrors", "-1", 0 );
-		r_fastsky = ri.Cvar_Get( "r_fastsky", "0", CVAR_ARCHIVE );
-		r_drawSun = ri.Cvar_Get( "r_drawSun", "0", 0 );
-		r_finish = ri.Cvar_Get( "r_finish", "0", CVAR_CHEAT );
-		r_textureMode = ri.Cvar_Get( "r_textureMode", "GL_LINEAR_MIPMAP_LINEAR", CVAR_ARCHIVE );
-		r_swapInterval = ri.Cvar_Get( "r_swapInterval", "0", CVAR_ARCHIVE );
-		r_gamma = ri.Cvar_Get( "r_gamma", "1.0", CVAR_ARCHIVE );
-		r_facePlaneCull = ri.Cvar_Get( "r_facePlaneCull", "1", 0 );
-
-		r_ambientScale = ri.Cvar_Get( "r_ambientScale", "1.0", CVAR_CHEAT | CVAR_LATCH );
-		r_lightScale = ri.Cvar_Get( "r_lightScale", "2", CVAR_CHEAT );
-
-		r_vboFaces = ri.Cvar_Get( "r_vboFaces", "1", CVAR_CHEAT );
-		r_vboCurves = ri.Cvar_Get( "r_vboCurves", "1", CVAR_CHEAT );
-		r_vboTriangles = ri.Cvar_Get( "r_vboTriangles", "1", CVAR_CHEAT );
-		r_vboShadows = ri.Cvar_Get( "r_vboShadows", "1", CVAR_CHEAT );
-		r_vboLighting = ri.Cvar_Get( "r_vboLighting", "1", CVAR_CHEAT );
-		r_vboModels = ri.Cvar_Get( "r_vboModels", "1", CVAR_LATCH );
-		r_vboVertexSkinning = ri.Cvar_Get( "r_vboVertexSkinning", "1",  CVAR_LATCH );
-		r_vboDeformVertexes = ri.Cvar_Get( "r_vboDeformVertexes", "1",  CVAR_LATCH );
-
-		r_mergeLeafSurfaces = ri.Cvar_Get( "r_mergeLeafSurfaces", "1",  CVAR_LATCH );
-
-		r_evsmPostProcess = ri.Cvar_Get( "r_evsmPostProcess", "0",  CVAR_LATCH );
-
-		r_printShaders = ri.Cvar_Get( "r_printShaders", "0", 0 );
-
-		r_bloom = ri.Cvar_Get( "r_bloom", "0", CVAR_LATCH | CVAR_ARCHIVE );
-		r_bloomBlur = ri.Cvar_Get( "r_bloomBlur", "1.0", CVAR_CHEAT );
-		r_bloomPasses = ri.Cvar_Get( "r_bloomPasses", "2", CVAR_CHEAT );
-		r_FXAA = ri.Cvar_Get( "r_FXAA", "0", CVAR_LATCH | CVAR_ARCHIVE );
-		r_ssao = ri.Cvar_Get( "r_ssao", "0", CVAR_LATCH | CVAR_ARCHIVE );
-
-		// temporary variables that can change at any time
-		r_showImages = ri.Cvar_Get( "r_showImages", "0", CVAR_TEMP );
-
-		r_debugSort = ri.Cvar_Get( "r_debugSort", "0", CVAR_CHEAT );
-
-		r_nocurves = ri.Cvar_Get( "r_nocurves", "0", CVAR_CHEAT );
-		r_lightScissors = ri.Cvar_Get( "r_lightScissors", "1", CVAR_ARCHIVE );
-		AssertCvarRange( r_lightScissors, 0, 2, true );
-
-		r_noLightVisCull = ri.Cvar_Get( "r_noLightVisCull", "0", CVAR_CHEAT );
-		r_noInteractionSort = ri.Cvar_Get( "r_noInteractionSort", "0", CVAR_CHEAT );
-		r_dynamicLight = ri.Cvar_Get( "r_dynamicLight", "2", CVAR_LATCH | CVAR_ARCHIVE );
-
-		r_staticLight = ri.Cvar_Get( "r_staticLight", "2", CVAR_ARCHIVE );
-		r_drawworld = ri.Cvar_Get( "r_drawworld", "1", CVAR_CHEAT );
-		r_portalOnly = ri.Cvar_Get( "r_portalOnly", "0", CVAR_CHEAT );
-		r_portalSky = ri.Cvar_Get( "cg_skybox", "1", 0 );
-		r_max_portal_levels = ri.Cvar_Get( "r_max_portal_levels", "5", 0 );
-
-		r_showSmp = ri.Cvar_Get( "r_showSmp", "0", CVAR_CHEAT );
-		r_skipBackEnd = ri.Cvar_Get( "r_skipBackEnd", "0", CVAR_CHEAT );
-		r_skipLightBuffer = ri.Cvar_Get( "r_skipLightBuffer", "0", CVAR_CHEAT );
-
-		r_measureOverdraw = ri.Cvar_Get( "r_measureOverdraw", "0", CVAR_CHEAT );
-		r_lodScale = ri.Cvar_Get( "r_lodScale", "5", CVAR_CHEAT );
-		r_lodTest = ri.Cvar_Get( "r_lodTest", "0.5", CVAR_CHEAT );
-		r_norefresh = ri.Cvar_Get( "r_norefresh", "0", CVAR_CHEAT );
-		r_drawentities = ri.Cvar_Get( "r_drawentities", "1", CVAR_CHEAT );
-		r_drawpolies = ri.Cvar_Get( "r_drawpolies", "1", CVAR_CHEAT );
-		r_ignore = ri.Cvar_Get( "r_ignore", "1", CVAR_CHEAT );
-		r_nocull = ri.Cvar_Get( "r_nocull", "0", CVAR_CHEAT );
-		r_novis = ri.Cvar_Get( "r_novis", "0", CVAR_CHEAT );
-		r_showcluster = ri.Cvar_Get( "r_showcluster", "0", CVAR_CHEAT );
-		r_speeds = ri.Cvar_Get( "r_speeds", "0", 0 );
-		r_verbose = ri.Cvar_Get( "r_verbose", "0", CVAR_CHEAT );
-		r_logFile = ri.Cvar_Get( "r_logFile", "0", CVAR_CHEAT );
-		r_debugSurface = ri.Cvar_Get( "r_debugSurface", "0", CVAR_CHEAT );
-		r_nobind = ri.Cvar_Get( "r_nobind", "0", CVAR_CHEAT );
-		r_clear = ri.Cvar_Get( "r_clear", "1", 0 );
-		r_offsetFactor = ri.Cvar_Get( "r_offsetFactor", "-1", CVAR_CHEAT );
-		r_offsetUnits = ri.Cvar_Get( "r_offsetUnits", "-2", CVAR_CHEAT );
-
-		r_physicalMapping = ri.Cvar_Get( "r_physicalMapping", "1", CVAR_LATCH | CVAR_ARCHIVE );
-		r_specularExponentMin = ri.Cvar_Get( "r_specularExponentMin", "0", CVAR_CHEAT );
-		r_specularExponentMax = ri.Cvar_Get( "r_specularExponentMax", "16", CVAR_CHEAT );
-		r_specularScale = ri.Cvar_Get( "r_specularScale", "1.0", CVAR_CHEAT | CVAR_LATCH );
-		r_specularMapping = ri.Cvar_Get( "r_specularMapping", "1", CVAR_LATCH | CVAR_ARCHIVE );
-		r_deluxeMapping = ri.Cvar_Get( "r_deluxeMapping", "1", CVAR_LATCH | CVAR_ARCHIVE );
-		r_normalScale = ri.Cvar_Get( "r_normalScale", "1.0", CVAR_ARCHIVE );
-		r_normalMapping = ri.Cvar_Get( "r_normalMapping", "1", CVAR_LATCH | CVAR_ARCHIVE );
-		r_highQualityNormalMapping = ri.Cvar_Get( "r_highQualityNormalMapping", "0",  CVAR_LATCH );
-		r_liquidMapping = ri.Cvar_Get( "r_liquidMapping", "0", CVAR_LATCH | CVAR_ARCHIVE );
-		r_reliefDepthScale = ri.Cvar_Get( "r_reliefDepthScale", "0.03", CVAR_CHEAT );
-		r_reliefMapping = ri.Cvar_Get( "r_reliefMapping", "0", CVAR_LATCH | CVAR_ARCHIVE );
-		r_glowMapping = ri.Cvar_Get( "r_glowMapping", "1", CVAR_LATCH );
-		r_reflectionMapping = ri.Cvar_Get( "r_reflectionMapping", "0", CVAR_LATCH | CVAR_ARCHIVE );
-
-		r_wrapAroundLighting = ri.Cvar_Get( "r_wrapAroundLighting", "0.7", CVAR_CHEAT | CVAR_LATCH );
-		r_halfLambertLighting = ri.Cvar_Get( "r_halfLambertLighting", "1", CVAR_LATCH | CVAR_ARCHIVE );
-		r_rimLighting = ri.Cvar_Get( "r_rimLighting", "0",  CVAR_LATCH | CVAR_ARCHIVE );
-		r_rimExponent = ri.Cvar_Get( "r_rimExponent", "3", CVAR_CHEAT | CVAR_LATCH );
-		AssertCvarRange( r_rimExponent, 0.5, 8.0, false );
-
-		r_drawBuffer = ri.Cvar_Get( "r_drawBuffer", "GL_BACK", CVAR_CHEAT );
-		r_lockpvs = ri.Cvar_Get( "r_lockpvs", "0", CVAR_CHEAT );
-		r_noportals = ri.Cvar_Get( "r_noportals", "0", CVAR_CHEAT );
-
-		r_shadows = ri.Cvar_Get( "cg_shadows", "1",  CVAR_LATCH );
-		AssertCvarRange( r_shadows, 0, Util::ordinal(shadowingMode_t::SHADOWING_EVSM32), true );
-
-		r_softShadows = ri.Cvar_Get( "r_softShadows", "0",  CVAR_LATCH );
-		AssertCvarRange( r_softShadows, 0, 6, true );
-
-		r_softShadowsPP = ri.Cvar_Get( "r_softShadowsPP", "0",  CVAR_LATCH );
-
-		r_shadowBlur = ri.Cvar_Get( "r_shadowBlur", "2",  CVAR_LATCH );
-
-		r_shadowMapQuality = ri.Cvar_Get( "r_shadowMapQuality", "3",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapQuality, 0, 4, true );
-
-		r_shadowMapSizeUltra = ri.Cvar_Get( "r_shadowMapSizeUltra", "1024",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeUltra, 32, 2048, true );
-
-		r_shadowMapSizeVeryHigh = ri.Cvar_Get( "r_shadowMapSizeVeryHigh", "512",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeVeryHigh, 32, 2048, true );
-
-		r_shadowMapSizeHigh = ri.Cvar_Get( "r_shadowMapSizeHigh", "256",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeHigh, 32, 2048, true );
-
-		r_shadowMapSizeMedium = ri.Cvar_Get( "r_shadowMapSizeMedium", "128",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeMedium, 32, 2048, true );
-
-		r_shadowMapSizeLow = ri.Cvar_Get( "r_shadowMapSizeLow", "64",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeLow, 32, 2048, true );
-
 		shadowMapResolutions[ 0 ] = r_shadowMapSizeUltra->integer;
 		shadowMapResolutions[ 1 ] = r_shadowMapSizeVeryHigh->integer;
 		shadowMapResolutions[ 2 ] = r_shadowMapSizeHigh->integer;
 		shadowMapResolutions[ 3 ] = r_shadowMapSizeMedium->integer;
 		shadowMapResolutions[ 4 ] = r_shadowMapSizeLow->integer;
 
-		r_shadowMapSizeSunUltra = ri.Cvar_Get( "r_shadowMapSizeSunUltra", "1024",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeSunUltra, 32, 2048, true );
-
-		r_shadowMapSizeSunVeryHigh = ri.Cvar_Get( "r_shadowMapSizeSunVeryHigh", "1024",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeSunVeryHigh, 512, 2048, true );
-
-		r_shadowMapSizeSunHigh = ri.Cvar_Get( "r_shadowMapSizeSunHigh", "1024",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeSunHigh, 512, 2048, true );
-
-		r_shadowMapSizeSunMedium = ri.Cvar_Get( "r_shadowMapSizeSunMedium", "1024",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeSunMedium, 512, 2048, true );
-
-		r_shadowMapSizeSunLow = ri.Cvar_Get( "r_shadowMapSizeSunLow", "1024",  CVAR_LATCH );
-		AssertCvarRange( r_shadowMapSizeSunLow, 512, 2048, true );
-
 		sunShadowMapResolutions[ 0 ] = r_shadowMapSizeSunUltra->integer;
 		sunShadowMapResolutions[ 1 ] = r_shadowMapSizeSunVeryHigh->integer;
 		sunShadowMapResolutions[ 2 ] = r_shadowMapSizeSunHigh->integer;
 		sunShadowMapResolutions[ 3 ] = r_shadowMapSizeSunMedium->integer;
 		sunShadowMapResolutions[ 4 ] = r_shadowMapSizeSunLow->integer;
-
-		r_shadowOffsetFactor = ri.Cvar_Get( "r_shadowOffsetFactor", "0", CVAR_CHEAT );
-		r_shadowOffsetUnits = ri.Cvar_Get( "r_shadowOffsetUnits", "0", CVAR_CHEAT );
-		r_shadowLodBias = ri.Cvar_Get( "r_shadowLodBias", "0", CVAR_CHEAT );
-		r_shadowLodScale = ri.Cvar_Get( "r_shadowLodScale", "0.8", CVAR_CHEAT );
-		r_noShadowPyramids = ri.Cvar_Get( "r_noShadowPyramids", "0", CVAR_CHEAT );
-		r_cullShadowPyramidFaces = ri.Cvar_Get( "r_cullShadowPyramidFaces", "0", CVAR_CHEAT );
-		r_cullShadowPyramidCurves = ri.Cvar_Get( "r_cullShadowPyramidCurves", "1", CVAR_CHEAT );
-		r_cullShadowPyramidTriangles = ri.Cvar_Get( "r_cullShadowPyramidTriangles", "1", CVAR_CHEAT );
-		r_noShadowFrustums = ri.Cvar_Get( "r_noShadowFrustums", "0", CVAR_CHEAT );
-		r_noLightFrustums = ri.Cvar_Get( "r_noLightFrustums", "1", CVAR_CHEAT );
-
-		r_maxPolys = ri.Cvar_Get( "r_maxpolys", "10000", CVAR_LATCH );  // 600 in vanilla Q3A
-		AssertCvarRange( r_maxPolys, 600, 30000, true );
-
-		r_maxPolyVerts = ri.Cvar_Get( "r_maxpolyverts", "100000", CVAR_LATCH );  // 3000 in vanilla Q3A
-		AssertCvarRange( r_maxPolyVerts, 3000, 200000, true );
-
-		r_showTris = ri.Cvar_Get( "r_showTris", "0", CVAR_CHEAT );
-		r_showSky = ri.Cvar_Get( "r_showSky", "0", CVAR_CHEAT );
-		r_showShadowVolumes = ri.Cvar_Get( "r_showShadowVolumes", "0", CVAR_CHEAT );
-		r_showShadowLod = ri.Cvar_Get( "r_showShadowLod", "0", CVAR_CHEAT );
-		r_showShadowMaps = ri.Cvar_Get( "r_showShadowMaps", "0", CVAR_CHEAT );
-		r_showSkeleton = ri.Cvar_Get( "r_showSkeleton", "0", CVAR_CHEAT );
-		r_showEntityTransforms = ri.Cvar_Get( "r_showEntityTransforms", "0", CVAR_CHEAT );
-		r_showLightTransforms = ri.Cvar_Get( "r_showLightTransforms", "0", CVAR_CHEAT );
-		r_showLightInteractions = ri.Cvar_Get( "r_showLightInteractions", "0", CVAR_CHEAT );
-		r_showLightScissors = ri.Cvar_Get( "r_showLightScissors", "0", CVAR_CHEAT );
-		r_showLightBatches = ri.Cvar_Get( "r_showLightBatches", "0", CVAR_CHEAT );
-		r_showLightGrid = ri.Cvar_Get( "r_showLightGrid", "0", CVAR_CHEAT );
-		r_showLightTiles = ri.Cvar_Get("r_showLightTiles", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_showBatches = ri.Cvar_Get( "r_showBatches", "0", CVAR_CHEAT );
-		r_showLightMaps = ri.Cvar_Get( "r_showLightMaps", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_showDeluxeMaps = ri.Cvar_Get( "r_showDeluxeMaps", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_showNormalMaps = ri.Cvar_Get( "r_showNormalMaps", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_showMaterialMaps = ri.Cvar_Get( "r_showMaterialMaps", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_showAreaPortals = ri.Cvar_Get( "r_showAreaPortals", "0", CVAR_CHEAT );
-		r_showCubeProbes = ri.Cvar_Get( "r_showCubeProbes", "0", CVAR_CHEAT );
-		r_showBspNodes = ri.Cvar_Get( "r_showBspNodes", "0", CVAR_CHEAT );
-		r_showParallelShadowSplits = ri.Cvar_Get( "r_showParallelShadowSplits", "0", CVAR_CHEAT | CVAR_LATCH );
-		r_showDecalProjectors = ri.Cvar_Get( "r_showDecalProjectors", "0", CVAR_CHEAT );
-
-		r_fontScale = ri.Cvar_Get( "r_fontScale", "36",  CVAR_LATCH );
 
 		// make sure all the commands added here are also removed in R_Shutdown
 		ri.Cmd_AddCommand( "imagelist", R_ImageList_f );
