+++ src/engine/sys/sdl_glimp.cpp
@@ -33,7 +33,8 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include "framework/CommandSystem.h"
 #include "framework/CvarSystem.h"
 
-static Log::Logger logger("glconfig", "", Log::Level::NOTICE);
+extern Log::Logger glConfigLogger;
+#define logger glConfigLogger
 
 SDL_Window         *window = nullptr;
 static SDL_GLContext glContext = nullptr;
@@ -862,296 +863,6 @@ static bool GLimp_StartDriverAndSetMode( int mode, bool fullscreen, bool noborde
 	return true;
 }
 
-static GLenum debugTypes[] =
-{
-	0,
-	GL_DEBUG_TYPE_ERROR_ARB,
-	GL_DEBUG_TYPE_DEPRECATED_BEHAVIOR_ARB,
-	GL_DEBUG_TYPE_UNDEFINED_BEHAVIOR_ARB,
-	GL_DEBUG_TYPE_PORTABILITY_ARB,
-	GL_DEBUG_TYPE_PERFORMANCE_ARB,
-	GL_DEBUG_TYPE_OTHER_ARB
-};
-
-#ifdef _WIN32
-#define DEBUG_CALLBACK_CALL APIENTRY
-#else
-#define DEBUG_CALLBACK_CALL
-#endif
-static void DEBUG_CALLBACK_CALL GLimp_DebugCallback( GLenum, GLenum type, GLuint,
-                                       GLenum severity, GLsizei, const GLchar *message, const void* )
-{
-	const char *debugTypeName;
-	const char *debugSeverity;
-
-	if ( r_glDebugMode->integer <= Util::ordinal(glDebugModes_t::GLDEBUG_NONE))
-	{
-		return;
-	}
-
-	if ( r_glDebugMode->integer < Util::ordinal(glDebugModes_t::GLDEBUG_ALL))
-	{
-		if ( debugTypes[ r_glDebugMode->integer ] != type )
-		{
-			return;
-		}
-	}
-
-	switch ( type )
-	{
-		case GL_DEBUG_TYPE_ERROR_ARB:
-			debugTypeName = "DEBUG_TYPE_ERROR";
-			break;
-		case GL_DEBUG_TYPE_DEPRECATED_BEHAVIOR_ARB:
-			debugTypeName = "DEBUG_TYPE_DEPRECATED_BEHAVIOR";
-			break;
-		case GL_DEBUG_TYPE_UNDEFINED_BEHAVIOR_ARB:
-			debugTypeName = "DEBUG_TYPE_UNDEFINED_BEHAVIOR";
-			break;
-		case GL_DEBUG_TYPE_PORTABILITY_ARB:
-			debugTypeName = "DEBUG_TYPE_PORTABILITY";
-			break;
-		case GL_DEBUG_TYPE_PERFORMANCE_ARB:
-			debugTypeName = "DEBUG_TYPE_PERFORMANCE";
-			break;
-		case GL_DEBUG_TYPE_OTHER_ARB:
-			debugTypeName = "DEBUG_TYPE_OTHER";
-			break;
-		default:
-			debugTypeName = "DEBUG_TYPE_UNKNOWN";
-			break;
-	}
-
-	switch ( severity )
-	{
-		case GL_DEBUG_SEVERITY_HIGH_ARB:
-			debugSeverity = "high";
-			break;
-		case GL_DEBUG_SEVERITY_MEDIUM_ARB:
-			debugSeverity = "med";
-			break;
-		case GL_DEBUG_SEVERITY_LOW_ARB:
-			debugSeverity = "low";
-			break;
-		default:
-			debugSeverity = "none";
-			break;
-	}
-
-	logger.Warn("%s: severity: %s msg: %s", debugTypeName, debugSeverity, message );
-}
-
-/*
-===============
-GLimp_InitExtensions
-===============
-*/
-
-/* ExtFlag_CORE means the extension is known to be an OpenGL 3 core extension.
-The code considers the extension is available even if the extension is not listed
-if the driver pretends to support OpenGL Core 3 and we know this extension is part
-of OpenGL Core 3. */
-
-enum {
-	ExtFlag_NONE,
-	ExtFlag_REQUIRED = BIT( 1 ),
-	ExtFlag_CORE = BIT( 2 ),
-};
-
-static bool LoadExt( int flags, bool hasExt, const char* name, bool test = true )
-{
-	if ( hasExt || ( flags & ExtFlag_CORE && glConfig2.glCoreProfile) )
-	{
-		if ( test )
-		{
-			logger.WithoutSuppression().Notice( "...using GL_%s", name );
-			return true;
-		}
-		else
-		{
-			// Required extension can't be made optional
-			ASSERT( !( flags & ExtFlag_REQUIRED ) );
-
-			logger.WithoutSuppression().Notice( "...ignoring GL_%s", name );
-		}
-	}
-	else
-	{
-		if ( flags & ExtFlag_REQUIRED )
-		{
-			Sys::Error( "Required extension GL_%s is missing", name );
-		}
-		else
-		{
-			logger.WithoutSuppression().Notice( "...GL_%s not found", name );
-		}
-	}
-	return false;
-}
-
-#define LOAD_EXTENSION(flags, ext) LoadExt(flags, GLEW_##ext, #ext)
-
-#define LOAD_EXTENSION_WITH_TEST(flags, ext, test) LoadExt(flags, GLEW_##ext, #ext, test)
-
-static void GLimp_InitExtensions()
-{
-	logger.Notice("Initializing OpenGL extensions" );
-
-	if ( LOAD_EXTENSION_WITH_TEST( ExtFlag_NONE, ARB_debug_output, r_glDebugProfile->value ) )
-	{
-		glDebugMessageCallbackARB( (GLDEBUGPROCARB)GLimp_DebugCallback, nullptr );
-		glEnable( GL_DEBUG_OUTPUT_SYNCHRONOUS_ARB );
-	}
-
-	// Shader limits
-	glGetIntegerv( GL_MAX_VERTEX_UNIFORM_COMPONENTS_ARB, &glConfig2.maxVertexUniforms );
-	glGetIntegerv( GL_MAX_VERTEX_ATTRIBS_ARB, &glConfig2.maxVertexAttribs );
-
-	int reservedComponents = 36 * 10; // approximation how many uniforms we have besides the bone matrices
-	glConfig2.maxVertexSkinningBones = Math::Clamp( ( glConfig2.maxVertexUniforms - reservedComponents ) / 16, 0, MAX_BONES );
-	glConfig2.vboVertexSkinningAvailable = r_vboVertexSkinning->integer && ( ( glConfig2.maxVertexSkinningBones >= 12 ) ? true : false );
-
-	// GLSL
-
-	Q_strncpyz( glConfig2.shadingLanguageVersionString, ( char * ) glGetString( GL_SHADING_LANGUAGE_VERSION_ARB ),
-				sizeof( glConfig2.shadingLanguageVersionString ) );
-	int majorVersion, minorVersion;
-	if ( sscanf( glConfig2.shadingLanguageVersionString, "%i.%i", &majorVersion, &minorVersion ) != 2 )
-	{
-		logger.Warn("unrecognized shading language version string format" );
-	}
-	glConfig2.shadingLanguageVersion = majorVersion * 100 + minorVersion;
-
-	logger.Notice("...found shading language version %i", glConfig2.shadingLanguageVersion );
-
-	// Texture formats and compression
-	glGetIntegerv( GL_MAX_CUBE_MAP_TEXTURE_SIZE_ARB, &glConfig2.maxCubeMapTextureSize );
-
-	// made required in OpenGL 3.0
-	glConfig2.textureHalfFloatAvailable =  LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, ARB_half_float_pixel, r_ext_half_float_pixel->value );
-
-	// made required in OpenGL 3.0
-	glConfig2.textureFloatAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, ARB_texture_float, r_ext_texture_float->value );
-
-	// made required in OpenGL 3.0
-	glConfig2.gpuShader4Available = LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, EXT_gpu_shader4, r_ext_gpu_shader4->value );
-
-	// made required in OpenGL 3.0
-	// GL_EXT_texture_integer can be used in shaders only if GL_EXT_gpu_shader4 is also available
-	glConfig2.textureIntegerAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, EXT_texture_integer, r_ext_texture_integer->value )
-	  && glConfig2.gpuShader4Available;
-
-	// made required in OpenGL 3.0
-	glConfig2.textureRGAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, ARB_texture_rg, r_ext_texture_rg->value );
-
-	{
-		/* GT218-based GPU with Nvidia 340.108 driver advertising
-		ARB_texture_gather extension is know to fail to compile
-		the depthtile1 GLSL shader.
-
-		See https://github.com/DaemonEngine/Daemon/issues/368
-
-		Unfortunately this workaround may also disable the feature for
-		all GPUs using this driver even if we don't know if some of them
-		are not affected by the bug while advertising this extension, but
-		there is no known easy way to detect GT218-based cards. Not all cards
-		using 340 driver supports this extension anyway, like the G92 one.
-
-		We can assume cards not using the 340 driver are not GT218 ones and
-		are not affected.
-
-		Usually, those GT218 cards are not powerful enough for dynamic
-		lighting so it is likely this feature would be disabled to
-		get acceptable framerate on this hardware anyway, making the
-		need for such extension and the related shader code useless. */
-		bool foundNvidia340 = ( Q_stristr( glConfig.vendor_string, "NVIDIA Corporation" ) && Q_stristr( glConfig.version_string, "NVIDIA 340." ) );
-
-		if ( foundNvidia340 )
-		{
-			// No need for WithoutSuppression for something which can only be printed once per renderer restart.
-			logger.Notice("...found buggy Nvidia 340 driver");
-		}
-
-		// made required in OpenGL 4.0
-		glConfig2.textureGatherAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_NONE, ARB_texture_gather, r_arb_texture_gather->value && !foundNvidia340 );
-	}
-
-	// made required in OpenGL 1.3
-	glConfig.textureCompression = textureCompression_t::TC_NONE;
-	if( LOAD_EXTENSION( ExtFlag_NONE, EXT_texture_compression_s3tc ) )
-	{
-		glConfig.textureCompression = textureCompression_t::TC_S3TC;
-	}
-
-	// made required in OpenGL 3.0
-	glConfig2.textureCompressionRGTCAvailable = LOAD_EXTENSION( ExtFlag_CORE, ARB_texture_compression_rgtc );
-
-	// Texture - others
-	glConfig2.textureAnisotropyAvailable = false;
-	if ( LOAD_EXTENSION_WITH_TEST( ExtFlag_NONE, EXT_texture_filter_anisotropic, r_ext_texture_filter_anisotropic->value ) )
-	{
-		glGetFloatv( GL_MAX_TEXTURE_MAX_ANISOTROPY_EXT, &glConfig2.maxTextureAnisotropy );
-		glConfig2.textureAnisotropyAvailable = true;
-	}
-
-	// VAO and VBO
-	// made required in OpenGL 3.0
-
-	LOAD_EXTENSION( ExtFlag_REQUIRED | ExtFlag_CORE, ARB_half_float_vertex );
-
-	// made required in OpenGL 3.0
-	LOAD_EXTENSION( ExtFlag_REQUIRED | ExtFlag_CORE, ARB_framebuffer_object );
-
-	// FBO
-	glGetIntegerv( GL_MAX_RENDERBUFFER_SIZE, &glConfig2.maxRenderbufferSize );
-	glGetIntegerv( GL_MAX_COLOR_ATTACHMENTS, &glConfig2.maxColorAttachments );
-
-	// made required in OpenGL 1.5
-	glConfig2.occlusionQueryAvailable = false;
-	glConfig2.occlusionQueryBits = 0;
-	if ( r_ext_occlusion_query->integer != 0 )
-	{
-		glConfig2.occlusionQueryAvailable = true;
-		glGetQueryiv( GL_SAMPLES_PASSED, GL_QUERY_COUNTER_BITS, &glConfig2.occlusionQueryBits );
-	}
-
-	// made required in OpenGL 2.0
-	glConfig2.drawBuffersAvailable = false;
-	if ( r_ext_draw_buffers->integer != 0 )
-	{
-		glGetIntegerv( GL_MAX_DRAW_BUFFERS, &glConfig2.maxDrawBuffers );
-		glConfig2.drawBuffersAvailable = true;
-	}
-
-	{
-		int formats = 0;
-
-		glGetIntegerv( GL_NUM_PROGRAM_BINARY_FORMATS, &formats );
-
-		if ( formats == 0 )
-		{
-			// No need for WithoutSuppression for something which can only be printed once per renderer restart.
-			logger.Notice("...no program binary formats");
-		}
-
-		glConfig2.getProgramBinaryAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_NONE, ARB_get_program_binary, formats > 0 );
-	}
-
-	glConfig2.bufferStorageAvailable = false;
-	glConfig2.bufferStorageAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_NONE, ARB_buffer_storage, r_arb_buffer_storage->integer > 0 );
-
-	// made required since OpenGL 3.1
-	glConfig2.uniformBufferObjectAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, ARB_uniform_buffer_object, r_arb_uniform_buffer_object->value );
-
-	// made required in OpenGL 3.0
-	glConfig2.mapBufferRangeAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, ARB_map_buffer_range, r_arb_map_buffer_range->value );
-
-	// made required in OpenGL 3.2
-	glConfig2.syncAvailable = LOAD_EXTENSION_WITH_TEST( ExtFlag_CORE, ARB_sync, r_arb_sync->value );
-
-	GL_CheckErrors();
-}
-
 static const int R_MODE_FALLBACK = 3; // 640 * 480
 
 /* Support code for GLimp_Init */
@@ -1507,21 +1218,4 @@ void GLimp_HandleCvars()
 	}
 
 	// TODO: Update r_allowResize using SDL_SetWindowResizable when we have SDL 2.0.5
-}
-
-void GLimp_LogComment( const char *comment )
-{
-	static char buf[ 4096 ];
-
-	if ( r_logFile->integer && GLEW_ARB_debug_output )
-	{
-		// copy string and ensure it has a trailing '\0'
-		Q_strncpyz( buf, comment, sizeof( buf ) );
-
-		glDebugMessageInsertARB( GL_DEBUG_SOURCE_APPLICATION_ARB,
-					 GL_DEBUG_TYPE_OTHER_ARB,
-					 0,
-					 GL_DEBUG_SEVERITY_MEDIUM_ARB,
-					 strlen( buf ), buf );
-	}
 }
